restructure firmware-common
