+++ firmware-common/base/usb-interface.hpp
