--- firmware-common/robot2015/cpu/status.c
@@ -1,7 +1,0 @@
-#include "status.h"
-
-unsigned int robot_id;
-unsigned int failures;
-uint8_t motor_faults;
-uint8_t current_motor_faults;
-int base2008;
